{"schema":"decodex/commit/1","summary":"Extract overlay worker protocol owner","authority":"manual"}
