+++ packages/rsnap-overlay/src/worker.rs
@@ -1,3 +1,11 @@
+mod pending_requests;
+mod protocol;
+
+pub(crate) use self::protocol::{
+	CapturedMonitorRegionResponse, CapturedMonitorRegionResult, FreezeCaptureTarget,
+	WorkerErrorSource, WorkerRequest, WorkerRequestSendError, WorkerResponse,
+};
+
 #[cfg(all(test, target_os = "macos"))]
 use std::sync::atomic::{AtomicU64, Ordering};
 use std::sync::{
@@ -10,110 +18,13 @@ use std::time::{Duration, Instant};
 
 use image::RgbaImage;
 
+use self::pending_requests::PendingWorkerRequests;
 use crate::backend::CaptureBackend;
 use crate::png;
 #[cfg(not(target_os = "macos"))]
 use crate::state::LiveCursorSample;
 use crate::state::RectPoints;
-use crate::state::{GlobalPoint, MonitorRect, WindowHit, WindowListSnapshot};
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub(crate) enum FreezeCaptureTarget {
-	Monitor,
-	Window { window_id: u32 },
-}
-
-#[derive(Debug)]
-pub(crate) enum WorkerRequest {
-	HitTestWindow {
-		monitor: MonitorRect,
-		point: GlobalPoint,
-		request_id: u64,
-	},
-	#[cfg(not(target_os = "macos"))]
-	SampleLiveCursor {
-		monitor: MonitorRect,
-		point: GlobalPoint,
-		request_id: u64,
-		want_patch: bool,
-		patch_width_px: u32,
-		patch_height_px: u32,
-	},
-	RefreshWindowList,
-	FreezeCapture {
-		monitor: MonitorRect,
-		target: FreezeCaptureTarget,
-	},
-	CaptureMonitorRegion {
-		monitor: MonitorRect,
-		rect_px: RectPoints,
-		request_id: u64,
-	},
-	EncodePng {
-		image: RgbaImage,
-	},
-}
-
-#[derive(Debug)]
-pub(crate) enum WorkerResponse {
-	#[cfg(not(target_os = "macos"))]
-	SampledLiveCursor {
-		monitor: MonitorRect,
-		point: GlobalPoint,
-		request_id: u64,
-		sample: LiveCursorSample,
-	},
-	HitTestWindow {
-		monitor: MonitorRect,
-		point: GlobalPoint,
-		request_id: u64,
-		hit: Option<WindowHit>,
-	},
-	RefreshedWindowList {
-		snapshot: Arc<WindowListSnapshot>,
-	},
-	CapturedFreeze {
-		monitor: MonitorRect,
-		image: RgbaImage,
-		window_image: Option<RgbaImage>,
-		captured_window_id: Option<u32>,
-	},
-	EncodedPng {
-		png_bytes: Vec<u8>,
-	},
-	Error {
-		source: WorkerErrorSource,
-		message: String,
-	},
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub(crate) enum WorkerErrorSource {
-	EncodePng,
-	FreezeCapture,
-	RefreshWindowList,
-	CaptureMonitorRegion,
-}
-
-#[derive(Debug)]
-pub(crate) enum CapturedMonitorRegionResult {
-	Image(RgbaImage),
-	NoNewFrame,
-}
-
-#[derive(Debug)]
-pub(crate) enum WorkerRequestSendError {
-	Full,
-	Disconnected,
-}
-
-#[derive(Debug)]
-pub(crate) struct CapturedMonitorRegionResponse {
-	pub(crate) monitor: MonitorRect,
-	pub(crate) rect_px: RectPoints,
-	pub(crate) request_id: u64,
-	pub(crate) result: CapturedMonitorRegionResult,
-}
+use crate::state::{GlobalPoint, MonitorRect};
 
 pub(crate) struct OverlayWorker {
 	req_tx: SyncSender<WorkerRequest>,
@@ -478,112 +389,6 @@ impl OverlayWorker {
 			Ok(msg) => Some(msg),
 			Err(TryRecvError::Empty) | Err(TryRecvError::Disconnected) => None,
 		}
-	}
-}
-
-#[derive(Default)]
-struct PendingWorkerRequests {
-	last_hit_test: Option<(MonitorRect, GlobalPoint, u64)>,
-	#[cfg(not(target_os = "macos"))]
-	last_sample_cursor: Option<(MonitorRect, GlobalPoint, u64, bool, u32, u32)>,
-	last_refresh_window_list: bool,
-	last_freeze: Option<(MonitorRect, FreezeCaptureTarget)>,
-	last_capture_region: Option<(MonitorRect, RectPoints, u64)>,
-	last_encode: Option<RgbaImage>,
-}
-impl PendingWorkerRequests {
-	fn record(&mut self, request: WorkerRequest) {
-		match request {
-			WorkerRequest::HitTestWindow { monitor, point, request_id } => {
-				self.last_hit_test = Some((monitor, point, request_id));
-			},
-			#[cfg(not(target_os = "macos"))]
-			WorkerRequest::SampleLiveCursor {
-				monitor,
-				point,
-				request_id,
-				want_patch,
-				patch_width_px,
-				patch_height_px,
-			} => {
-				self.last_sample_cursor =
-					Some((monitor, point, request_id, want_patch, patch_width_px, patch_height_px));
-			},
-			WorkerRequest::RefreshWindowList => {
-				self.last_refresh_window_list = true;
-			},
-			WorkerRequest::FreezeCapture { monitor, target } => {
-				self.last_freeze = Some((monitor, target));
-			},
-			WorkerRequest::CaptureMonitorRegion { monitor, rect_px, request_id } => {
-				self.last_capture_region = Some((monitor, rect_px, request_id));
-			},
-			WorkerRequest::EncodePng { image } => {
-				self.last_encode = Some(image);
-			},
-		}
-	}
-
-	fn dispatch(
-		self,
-		backend: &mut dyn CaptureBackend,
-		resp_tx: &Sender<WorkerResponse>,
-		_region_capture_resp_tx: &Sender<CapturedMonitorRegionResponse>,
-		response_waker: Option<&(dyn Fn() + Send + Sync)>,
-	) {
-		let mut handled_high_priority = false;
-
-		if let Some(image) = self.last_encode {
-			OverlayWorker::handle_encode_request(resp_tx, response_waker, image);
-
-			handled_high_priority = true;
-		}
-
-		if handled_high_priority {
-			return;
-		}
-
-		if let Some((monitor, target)) = self.last_freeze {
-			OverlayWorker::handle_freeze_request(backend, resp_tx, response_waker, monitor, target);
-
-			return;
-		}
-		if let Some((monitor, rect_px, request_id)) = self.last_capture_region {
-			OverlayWorker::handle_capture_monitor_region_request(
-				backend,
-				resp_tx,
-				_region_capture_resp_tx,
-				response_waker,
-				monitor,
-				rect_px,
-				request_id,
-			);
-
-			return;
-		}
-
-		if self.last_refresh_window_list {
-			OverlayWorker::handle_refresh_window_list_request(backend, resp_tx, response_waker);
-		}
-
-		#[cfg(not(target_os = "macos"))]
-		if let Some((monitor, point, request_id, want_patch, patch_width_px, patch_height_px)) =
-			self.last_sample_cursor
-		{
-			OverlayWorker::handle_sample_cursor_request(
-				backend,
-				resp_tx,
-				response_waker,
-				(monitor, point, request_id, want_patch, patch_width_px, patch_height_px),
-			);
-		}
-
-		OverlayWorker::handle_hit_test_request(
-			backend,
-			resp_tx,
-			response_waker,
-			self.last_hit_test,
-		);
 	}
 }
